add location Details viw

+++ findingWater/View/locationDitels.swift
@@ -18,6 +18,7 @@ struct locationDitels: View {
                GridItem(.flexible()),
                GridItem(.flexible())
         ]
+    @ObservedObject var viewModel2 = CreateAccount()
     @ObservedObject var mapViewModel = LocationMapViewModel()
     var body: some View {
         ScrollView {
@@ -32,6 +33,9 @@ struct locationDitels: View {
                       
                           //  viewModel2.EditProducti(id: viewModel2.currentUser?.id  ?? "reewwedsewd", value: true ?? true)
                             //mapViewModel.createRestaurant(mdel: place.name)
+                        
+                        if viewModel2.userSession != nil  {
+                            viewModel2.currentUser?.value.toggle()
                             
                             let docData: [String: Any] = [
                                 // "uid": uid,
@@ -49,6 +53,9 @@ struct locationDitels: View {
                                     print("Document successfully written!")
                                 }
                             }
+                            
+                            viewModel2.EditProducti(id: viewModel2.currentUser?.id  ?? "reewwedsewd", value: viewModel2.currentUser?.value ?? true)
+                        }
                        // viewModel2.fetchUser()
                        // print(" viewModel2.1\(viewModel2.currentUser?.value)")
                         
@@ -63,26 +70,25 @@ struct locationDitels: View {
                             .clipShape(Circle())
                             .overlay(Circle().stroke(Color.black, lineWidth: 1))
                     }
+                    .disabled( viewModel2.currentUser?.value == true)
+              
                     
                     ForEach(mapViewModel.user, id: \.self) { item in
                         Button{
-                            //viewModel2.currentUser?.value.toggle()
-                           
-                            Firestore.firestore().collection(soures.name).document(item.id ?? "eereded").delete() { err in
+                            if viewModel2.userSession != nil &&   viewModel2.currentUser?.value == true {
+                                viewModel2.currentUser?.value.toggle()
+                                
+                                Firestore.firestore().collection(soures.name).document(item.id ?? "eereded").delete() { err in
                                     if let err = err {
                                         print("Error removing document: \(err)")
                                     } else {
                                         print("Document successfully removed!")
                                     }
                                 }
-                                //viewModel2.currentUser?.value.toggle()
                                 
+                                viewModel2.EditProducti(id: viewModel2.currentUser?.id  ?? "reewwedsewd", value:  viewModel2.currentUser?.value ?? false )
                                 
-                               // viewModel2.EditProducti(id: viewModel2.currentUser?.id  ?? "reewwedsewd", value: false ?? false)
-                            //viewModel2.fetchUser()
-                           // print(" viewModel2.\(viewModel2.currentUser?.value)")
-                                
-                            
+                            }
                             
                         } label: {
                             Image(systemName: "person.fill")
@@ -94,11 +100,17 @@ struct locationDitels: View {
                         
                     }
                     .onAppear{
-                       // mapViewModel.fetchData2(mdel:place.name)
+                      
                        // print(viewModel2.currentUser?.value)
                        // self.viewModel2.fetchUser()
                     }
                 
+            }
+            .onAppear{
+                mapViewModel.fetchData2(mdel:soures.name)
+                self.viewModel2.fetchUser()
+                viewModel2.userSession
+                print(viewModel2.currentUser?.value )
             }
         }
     }
